// src/components/TableComponents/StandardTableTypes.d.ts
type StickyTableConfigType = {
  /** The default for this is 1. If you need more than 1 sticky header, then use this prop. */
  header?: number
  /** The default for this is 1. If you need more than 1 sticky left column, then use this prop. */
  left?: number
  /** This should be set to 1 or more if the far right column has an action button in it. */
  right?: number
}
export type ConfigItemType<DataItem, StyleGeneric> = {
  /** Unique name for the column. Should be associated with the name that you are sorting on for the column. */
  name: SortColumnProps['propName']
  /** Unique label for the column headers. */
  label: SortColumnProps['label']
  /** Optionally add custom JSX. This will be rendered to the right of the column header. */
  columnHeaderSubContent?: React.ReactNode
  /** Used if you have multiple options to sort on within a column. */
  options?: SortColumnProps['options']
  /** The JSX that will be rendered in a cell. */
  cell: {
    className?: string | ((data: DataItem) => string)
    children: (data: DataItem, index?: number) => React.ReactNode
    /** Optional style overrides for a cell. */
    style?: StyleGeneric
  }
  /** Determines the static first column in mobile and also the column that will not be able to be removed in column customizations. */
  mainColumn?: boolean
  /** Used if the API requires a lowercase parameter to be passed in. */
  lowerCaseParam?: SortColumnProps['lowerCaseParam']
  /** Define the tooltip for a column. */
  tooltip?: SortColumnProps['tooltip']
  /** Class name for a column. */
  className?: string
  /** Prevent sorting for a column. */
  noSort?: boolean
  /** Column that has a single button. Used for display purposes in mobile. */
  isButton?: boolean
  /** Filter for a column. */
  filter?: SortColumnProps['filter']
  /** Style overrides for a column. */
  style?: StyleGeneric
}
// ...
export type NestedRowPropsType<
  DataItem,
  ConfigItem extends ConfigItemType<DataItem, Record<string, unknown>>
> = {
  /** data passed into the nested row */
  data: DataItem
  /** Table configuration that will handle column headers and rendering column children. */
  config: ConfigItem[]
  /** Unique key from our nested data objects. Used for `key` in our `map` functions. */
  nestedDataKey?: keyof DataItem
  /** This gets passed down to SortColumn's 'active' prop */
  sortBy: SortColumnProps['active']
  nestedRowProps?: {
    /** Function to get array of records which are associated with parent record */
    getNestedData: (key: DataItem[keyof DataItem]) => void
    /** Flag to determine if the nested data is being fetched */
    nestedLoading: {
      [key: string]: boolean
    }
    /** Array of nested rows data for selected parent record */
    nestedData: {
      [key: string]: DataItem[]
    }
  }
  customColumnProps?: {
    /** Array of user selected columns that we want displayed in the UI */
    selected: Array<ConfigItem['label']>
    /** Function to get array of selected columns saved by the user */
    callout: (selectedList: string[]) => void
    /** Array of all the columns that can be displayed in the UI */
    list: string[]
  }
  /** The table needs to know which columns and rows are sticky. This object is where we define that. */
  stickyTableConfig?: StickyTableConfigType
  /** Unique key from our data objects to identify total row. Value should be boolean */
  totalRowKey?: keyof DataItem
  /** Selected row for showing the nested list */
  selectedNestedRowData?: DataItem[keyof DataItem][]
  showStickyClasses: boolean
}